Stop guessing values for some parameters.
The script was filling the values for some parameters, not present in
PuppetDB response, with values of other parameters. This is highly
dependent on every user Nagios virtual resources.

This feature has been removed. This version may generate a different
configuration from older releases of the script, given the same response
from PuppetDB. If you were using this feature, you may want to ensure
the following parameters are present on your Nagios resources:

 * 'host_name' for 'nagios_host' resources
 * 'host_name' for 'nagios_hostextinfo' resources
 * 'contact_name' for 'nagios_contact' resources
 * 'contactgroup_name' for 'nagios_contactgroup' resources
 * 'command_name' for 'nagios_command' resources

Fixes #6